+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## CHANGELOG
 
-### v0.7.2
+### v0.7.x
 
 1. Migrated to uhppoted-lib `config` command implementation
+2. Reworked `get-events` to return the `first`, `last` and `current` event indices.
+3. Reworked `get-event`  to handle `first`, `last`, `current`, `next` and a valid index.
 
 ### v0.7.1